requirements.txt
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.2.2
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.23.3
pandas==1.5.0
paramiko==2.11.0
protobuf==3.20.1
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sshtunnel==0.4.0
Werkzeug==2.2.2
